requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
httplib2==0.31.0
idna==3.11
oauth2client==4.1.3
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
